Disable pixelize for large images (#32)
This PR fixes a problem when large resolution images are shown in the
report. So it does not affect Linebender usecases, because snapshots are
usually small, but e.g. @Korf-tms uses Kompari for comparing A4 pdf
renders so the problem is triggered.

The problem is that Kompari uses css flag "pixelize" in report when an
image is zoomed. The idea is to avoid filtering for zoomed images so a
user can see individual pixels. However when an image is bigger than the
window and image size still has to be reduced, then "pixelize" reduces
the quality of the output.

The fix is simple, we disable "pixelize" for large images. We now define
"large" image as an image that has both dimensions greater then 400.