Paralelization of dir diff computing and report generation (#34)
This PR parallelizes directory diff computing and report generating.

Its add `rayon` dependency to `kompari` and `kompari-html`, but it was
indirectly compiled anyway because of oxipng.

(I think that it starts to crystalize that current `kompari` create
needs to be splitted into two `lightweight-kompari-for-usage-in-tests`
and `common-utilities-for-postprocessing`, I will start a separate
thread for this).