Remove dependency on `image` in Kompari (#36)
This is for compile time and binary size reasons; it also helps to avoid
some insane situations where image could read snapshots encoded in the
wrong format.

This PR cut debug build time of just Kompari to 2.3s from 2.6s. When we
also remove `thiserror` (which I intend to do), that will be cut down
significantly further; thiserror is hiding the impact of image.